Update admClean.csv
Adding the calculated drone decibels at the given frequency bands. Used the 79db profile of the DJI Mavic Pro,and the divergence between bands seen in the M60 tank profile, and modeled that same difference in bands for the drone's 79db. The changes are modeled with the same parameters as the tank (ex: mic distance at 30 meters).

So: Change in tank decibels between 80-100Hz is --> 1-(60.9/69.8) --> 0.13

Applying that same difference to the drone at 80-100hz --> 79 + (1-(79*0.13))